% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_tools.R
\name{cart2sphere}
\alias{cart2sphere}
\title{Convert a Matrix of Cartesian Coordinates into Spherical Coordinates}
\usage{
cart2sphere(x)
}
\arguments{
\item{x}{A 3 column matrix with x, y and z coordinates.}
}
\value{
A 3 column matrix with radius, inclination and azimuth.
}
\description{
Convert a Matrix of Cartesian Coordinates into Spherical Coordinates
}
\examples{
x <- rbind(
  c(0, 0, 1),
  c(sqrt(8 / 9), 0, -1 / 3),
  c(-sqrt(2 / 9), sqrt(2 / 3), -1 / 3),
  c(-sqrt(2 / 9), -sqrt(2 / 3), -1 / 3)
)
cart2sphere(x)
}
