% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_tools.R
\name{sphere2latlong}
\alias{sphere2latlong}
\title{Convert a Matrix of Spherical Coordinates into an Spatial Object}
\usage{
sphere2latlong(x)
}
\arguments{
\item{x}{A 3 column matrix with radius, inclination and azimuth.}
}
\description{
Convert a Matrix of Spherical Coordinates into an Spatial Object
}
\examples{
x <- rbind(
  c(0, 0, 1),
  c(sqrt(8 / 9), 0, -1 / 3),
  c(-sqrt(2 / 9), sqrt(2 / 3), -1 / 3),
  c(-sqrt(2 / 9), -sqrt(2 / 3), -1 / 3)
)
x <- cart2sphere(x)
sphere2latlong(x)
}
